Code refactored & ConnectionState are now int values instead of an enum for it to be inherited.

+++ Iris.Core/EndPointConnection.cs
@@ -6,7 +6,7 @@ namespace Iris.Core
     public class EndPointConnection : IEndPointConnection
     {
         public string Id { get; } = RandUtils.RandomId();
-        public int ConnectionState { get; set; }
+        public byte ConnectionState { get; set; }
         
         public readonly IPEndPoint EndPoint;
