+++ Iris.Core/IEndPointConnection.cs
@@ -3,6 +3,6 @@ namespace Iris.Core
     public interface IEndPointConnection
     {
         string Id { get; }
-        int ConnectionState { get; set; }
+        byte ConnectionState { get; set; }
     }
 }
